[优化]：调整一下ajaxTestAll函数细节

+++ app/src/main/java/io/legado/app/help/http/StrResponse.kt
@@ -20,7 +20,7 @@ class StrResponse {
         private set
     var errorBody: ResponseBody? = null
         private set
-    var callTime = 0L
+    var callTime = 0
 
     constructor(rawResponse: Response, body: String?) {
         this.raw = rawResponse
@@ -47,7 +47,7 @@ class StrResponse {
         this.errorBody = errorBody
     }
 
-    fun putCallTime(callTime: Long) {
+    fun putCallTime(callTime: Int) {
         this.callTime = callTime
     }
     fun raw() = raw
